requirements.txt
pywin32>=306
python-dateutil>=2.9
tzlocal>=5.2
PyYAML>=6.0
jinja2>=3.1
rich>=13.7